nv8

+++ Domain/DTO/Expense/ExpenseResponseDTO.cs
@@ -1,6 +1,6 @@
 ﻿namespace Domain.DTO.Expense
 {
-    public record ExpenseResponseDTO(string description, double value)
+    public record ExpenseResponseDTO(string description, decimal value)
     {
     }
 }
